Retire the WebSocket/STOMP stack now that events ride Server-Sent Events
The STOMP broker, its subscription interceptor, and the websocket push
half of the event bridge are gone; the bridge now feeds the SSE stream
registry alone and is named for what it does. Everything that lived in
the websocket package but serves the SSE transport moved out of it: the
application-wide ObjectMapper configuration to the gateway root, and the
access manager, capabilities, wire DTOs, and the bridge itself to the
sse package. The unused websocket-era packaged-event mirror is deleted
outright, along with the /wsapps security rule and the allowed-origin
setting only the broker consumed.

The websocket starter is replaced by an explicit web starter. The web
stack was in fact arriving transitively through webprotege-ipc, not the
websocket starter as once assumed, but a web application should declare
its web framework directly rather than lean on a library's transitive
graph either way.

Completes the gateway half of protegeproject/webprotege-gwt-ui#308
(epic protegeproject/webprotege-gwt-ui#303).
