+++ src/main/java/edu/stanford/protege/webprotege/gateway/SecurityConfig.java
@@ -63,7 +63,6 @@ public class SecurityConfig {
     public SecurityFilterChain resourceServerFilterChain(HttpSecurity http) throws Exception {
         http.csrf(AbstractHttpConfigurer::disable)
                 .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/wsapps").permitAll()
                         // The SSE stream cannot be bearer-authenticated (EventSource sends no Authorization
                         // header); it is guarded instead by the short-lived stream ticket that the controller
                         // redeems and re-checks for VIEW_PROJECT on every connect. The ticket-issuing endpoint
