+++ src/main/java/edu/stanford/protege/webprotege/gateway/sse/StreamTicketService.java
@@ -4,8 +4,6 @@ import edu.stanford.protege.webprotege.authorization.ProjectResource;
 import edu.stanford.protege.webprotege.authorization.Subject;
 import edu.stanford.protege.webprotege.common.ProjectId;
 import edu.stanford.protege.webprotege.common.UserId;
-import edu.stanford.protege.webprotege.gateway.websocket.AccessManager;
-import edu.stanford.protege.webprotege.gateway.websocket.dto.BuiltInCapability;
 import edu.stanford.protege.webprotege.ipc.ExecutionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
