+++ src/main/resources/application.yaml
@@ -31,7 +31,6 @@ webprotege:
     event-subscribe: true
     # Each gateway instance pins its own long-lived SSE streams, so every instance must see every event; give each its own queue on the fanout exchange rather than competing on one shared queue.
     event-queue-per-instance: true
-  allowedOrigin: webprotege-local.edu
   sse:
     # Keep-alive interval; must stay below the nginx proxy_read_timeout (60s) so idle streams survive.
     heartbeat-interval: 20s
