+++ src/test/java/edu/stanford/protege/webprotege/gateway/sse/HistoryReplaySseCatchUpServiceTest.java
@@ -6,10 +6,7 @@ import com.google.common.util.concurrent.MoreExecutors;
 import edu.stanford.protege.webprotege.common.ProjectId;
 import edu.stanford.protege.webprotege.common.UserId;
 import edu.stanford.protege.webprotege.event.EventTag;
-import edu.stanford.protege.webprotege.gateway.websocket.config.ObjectMapperConfiguration;
-import edu.stanford.protege.webprotege.gateway.websocket.dto.EventList;
-import edu.stanford.protege.webprotege.gateway.websocket.dto.ProjectEventsQueryRequest;
-import edu.stanford.protege.webprotege.gateway.websocket.dto.ProjectEventsQueryResponse;
+import edu.stanford.protege.webprotege.gateway.ObjectMapperConfiguration;
 import edu.stanford.protege.webprotege.ipc.CommandExecutor;
 import edu.stanford.protege.webprotege.ipc.ExecutionContext;
 import org.junit.jupiter.api.BeforeEach;
