+++ src/test/java/edu/stanford/protege/webprotege/gateway/sse/SseStreamRegistryTest.java
@@ -5,9 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import edu.stanford.protege.webprotege.common.ProjectId;
 import edu.stanford.protege.webprotege.common.UserId;
 import edu.stanford.protege.webprotege.event.EventTag;
-import edu.stanford.protege.webprotege.gateway.websocket.config.ObjectMapperConfiguration;
-import edu.stanford.protege.webprotege.gateway.websocket.dto.EventList;
-import edu.stanford.protege.webprotege.gateway.websocket.dto.ProjectEventsQueryResponse;
+import edu.stanford.protege.webprotege.gateway.ObjectMapperConfiguration;
 import edu.stanford.protege.webprotege.ipc.ExecutionContext;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
