+++ src/test/resources/application.yaml
@@ -25,7 +25,6 @@ webprotege:
   gateway:
     reply-channel: ${spring.application.name}-replies
     timeout: 600000 # Ten minutes
-  allowedOrigin: webprotege-local.edu
   minio:
     bucket-name: foobucket
     end-point: https://localhost:9000
